Add a minimal OgreBites example
Immediately setting up CEGUI and Ogre at the same time proved difficult. So as a compromise I made an easier setup using OgreBites. CEGUI may be introduced later should it be needed.

The current scene shows a sphere and a line, which is a good start for rendering a skeleton.